requirements.txt
streamlit==1.28.1
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
scikit-learn==1.3.2
pillow==10.0.1